+++ lib/src/random.dart
@@ -23,7 +23,7 @@ Uint32List randomNumbers(
 }) =>
     HashlibRandom(generator).nextNumbers(length);
 
-/// Fill the [buffer] with random numbers.
+/// Fills the [buffer] with random numbers.
 ///
 /// Both the [start] and [length] are in bytes.
 @pragma('vm:prefer-inline')
@@ -34,6 +34,23 @@ void fillRandom(
   RandomGenerator generator = RandomGenerator.system,
 }) =>
     HashlibRandom(generator).fill(buffer, start, length);
+
+/// Fills the [list] with random 32-bit numbers.
+///
+/// Both the [start] and [length] are in bytes.
+void fillNumbers(
+  List<int> list, {
+  int start = 0,
+  int? length,
+  RandomGenerator generator = RandomGenerator.system,
+}) {
+  int n = length ?? list.length;
+  if (n == 0) return;
+  var rand = HashlibRandom(generator);
+  for (; n > 0 && start < list.length; ++start, --n) {
+    list[start] = rand.nextInt();
+  }
+}
 
 /// Generate a list of random ASCII string of size [length].
 ///